+++ example/lib/example_copy_paste_from_playground/app_theme.dart
@@ -79,7 +79,6 @@ abstract final class AppTheme {
       useSecondary: true,
       useTertiary: true,
       keepPrimary: true,
-      useExpressiveOnContainerColors: true,
     ),
     tones: FlexSchemeVariant.jolly.tones(Brightness.light),
     visualDensity: FlexColorScheme.comfortablePlatformDensity,